+++ R/loocv_glm.R
@@ -81,14 +81,18 @@ loocv.spglm <- function(object, cv_predict = FALSE, se.fit = FALSE, local, ...) 
     # get w for later
     w <- fitted(object, type = "link")
 
+    extra_randcov_list <- get_extra_randcov_list(object, object$obdata, newdata = object$obdata)
+    extra_partition_list <- get_extra_partition_list(object, object$obdata, newdata = object$obdata)
+
     if (local_list$parallel) {
       # turn of parallel as it is used different in predict
       local_list$parallel <- FALSE
       cl <- parallel::makeCluster(local_list$ncores)
-      cv_predict_val_list <- parallel::parLapply(cl, seq_len(object$n), loocv_local_glm, object, se.fit, local_list)
+      cv_predict_val_list <- parallel::parLapply(cl, seq_len(object$n), loocv_local_glm, object, se.fit, local_list, extra_randcov_list = extra_randcov_list, extra_partition_list = extra_partition_list)
       cl <- parallel::stopCluster(cl)
     } else {
-      cv_predict_val_list <- lapply(seq_len(object$n), loocv_local_glm, object, se.fit, local_list)
+      cv_predict_val_list <- lapply(seq_len(object$n), loocv_local_glm, object, se.fit, local_list,
+                                    extra_randcov_list = extra_randcov_list, extra_partition_list = extra_partition_list)
     }
     if (se.fit) {
       cv_predict_val <- vapply(cv_predict_val_list, function(x) x$fit, numeric(1))
@@ -100,20 +104,49 @@ loocv.spglm <- function(object, cv_predict = FALSE, se.fit = FALSE, local, ...) 
 
   cv_predict_val_invlink <- invlink(cv_predict_val, object$family, object$size)
 
-  if (cv_predict) {
-    if (se.fit) {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val), se.fit = as.vector(cv_predict_se))
-    } else {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val))
-    }
+
+  cv_predict_error <- y - cv_predict_val_invlink
+  bias <- mean(cv_predict_error)
+  MSPE <- mean((cv_predict_error)^2)
+  RMSPE <- sqrt(MSPE)
+
+  loocv_stats <- tibble(
+    bias = bias,
+    MSPE = MSPE,
+    RMSPE = RMSPE
+  )
+
+  if (!cv_predict && ! se.fit) {
+    return(loocv_stats)
   } else {
-    if (se.fit) {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), se.fit = as.vector(cv_predict_se))
-    } else {
-      cv_output <- mean((cv_predict_val_invlink - y)^2)
+    loocv_out <- list()
+    loocv_out$stats <- loocv_stats
+
+    if (cv_predict) {
+      loocv_out$cv_predict <- cv_predict_val
     }
+
+    if (se.fit) {
+      loocv_out$se.fit <- as.vector(cv_predict_se)
+    }
+    return(loocv_out)
   }
-  cv_output
+#
+#
+#   if (cv_predict) {
+#     if (se.fit) {
+#       cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val), se.fit = as.vector(cv_predict_se))
+#     } else {
+#       cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val))
+#     }
+#   } else {
+#     if (se.fit) {
+#       cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), se.fit = as.vector(cv_predict_se))
+#     } else {
+#       cv_output <- mean((cv_predict_val_invlink - y)^2)
+#     }
+#   }
+#   cv_output
 }
 
 #' @rdname loocv
@@ -184,25 +217,70 @@ loocv.spgautor <- function(object, cv_predict = FALSE, se.fit = FALSE, local, ..
 
   cv_predict_val_invlink <- invlink(cv_predict_val, object$family, object$size)
 
-  if (cv_predict) {
-    if (se.fit) {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val), se.fit = as.vector(cv_predict_se))
-    } else {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val))
-    }
+  cv_predict_error <- y - cv_predict_val_invlink
+  bias <- mean(cv_predict_error)
+  MSPE <- mean((cv_predict_error)^2)
+  RMSPE <- sqrt(MSPE)
+
+  loocv_stats <- tibble(
+    bias = bias,
+    MSPE = MSPE,
+    RMSPE = RMSPE
+  )
+
+  if (!cv_predict && ! se.fit) {
+    return(loocv_stats)
   } else {
-    if (se.fit) {
-      cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), se.fit = as.vector(cv_predict_se))
-    } else {
-      cv_output <- mean((cv_predict_val_invlink - y)^2)
+    loocv_out <- list()
+    loocv_out$stats <- loocv_stats
+
+    if (cv_predict) {
+      loocv_out$cv_predict <- cv_predict_val
     }
+
+    if (se.fit) {
+      loocv_out$se.fit <- as.vector(cv_predict_se)
+    }
+    return(loocv_out)
   }
-  cv_output
+
+  # if (cv_predict) {
+  #   if (se.fit) {
+  #     cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val), se.fit = as.vector(cv_predict_se))
+  #   } else {
+  #     cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), cv_predict = as.vector(cv_predict_val))
+  #   }
+  # } else {
+  #   if (se.fit) {
+  #     cv_output <- list(mspe = mean((cv_predict_val_invlink - y)^2), se.fit = as.vector(cv_predict_se))
+  #   } else {
+  #     cv_output <- mean((cv_predict_val_invlink - y)^2)
+  #   }
+  # }
+  # cv_output
 }
 
-loocv_local_glm <- function(row, object, se.fit, local_list) {
+loocv_local_glm <- function(row, object, se.fit, local_list,
+                            extra_randcov_list = NULL, extra_partition_list = NULL) {
   object$fitted$link <- object$fitted$link[-row] # w needs to be subset
   newdata <- object$obdata[row, , drop = FALSE]
   object$obdata <- object$obdata[-row, , drop = FALSE]
-  predict(object, newdata = newdata, se.fit = se.fit, local = local_list)
+  # this is all so the randcov and partition steps are not repeated for each iteration
+  if (!is.null(extra_randcov_list)) {
+    extra_randcov_list$Z_index_obdata_list <- lapply(extra_randcov_list$Z_index_obdata_list,
+                                                     function(x) {
+                                                       x$reform_bar2_vals <- x$reform_bar2_vals[-row]
+                                                       x
+                                                     })
+  }
+
+  if (!is.null(extra_partition_list)) {
+    extra_partition_list$partition_index_obdata$reform_bar2_vals <- extra_partition_list$partition_index_obdata$reform_bar2_vals[-row]
+  }
+
+  predict(object, newdata = newdata, se.fit = se.fit, local = local_list,
+          extra_randcov_list = extra_randcov_list,
+          extra_partition_list = extra_partition_list
+  )
+
 }