+++ R/spmodel-package.R
@@ -6,7 +6,7 @@
 #' @importFrom graphics abline par points title
 #' @importFrom parallel detectCores makeCluster parLapply stopCluster
 #' @importFrom sf st_as_sf st_centroid st_coordinates st_crs st_drop_geometry st_intersects
-#' @importFrom stats AIC anova coef coefficients complete.cases confint cooks.distance
+#' @importFrom stats AIC anova coef coefficients complete.cases confint cooks.distance cor
 #'   dbeta dbinom dgamma dnbinom dpois
 #'   delete.response deviance dist fitted fitted.values formula .getXlevels hatvalues
 #'   influence kmeans lm logLik model.frame model.matrix model.offset model.response na.omit na.pass pchisq